requirements.txt
numba>=0.59
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
pyyaml>=6.0
openpyxl>=3.1
pyarrow>=14.0
requests>=2.31
